refactor(domain): move board files to subfolder
