+++ src/core/domain/board/BoardNode.ts
