+++ src/core/domain/board/BoardProjection.ts
