+++ src/core/domain/board/GridBoard.ts
