+++ src/core/domain/board/GridPosition.ts
