+++ src/core/usecases/commands/RemoveNode/RemoveNodeCommandHandler.ts
@@ -1,4 +1,4 @@
-import { type GridBoardRepository } from '../../../domain/GridBoardRepository';
+import { type GridBoardRepository } from '../../../domain/board/GridBoardRepository';
 import { RemoveNodeCommand } from './RemoveNodeCommand';
 
 export class RemoveNodeCommandHandler {
