+++ src/core/usecases/commands/UpdateNodeLabel/UpdateNodeLabelCommandHandler.ts
@@ -1,4 +1,4 @@
-import { type GridBoardRepository } from '../../../domain/GridBoardRepository';
+import { type GridBoardRepository } from '../../../domain/board/GridBoardRepository';
 import { UpdateNodeLabelCommand } from './UpdateNodeLabelCommand';
 
 export class UpdateNodeLabelCommandHandler {
