+++ src/ui/components/Slices/ScenarioDialog.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react';
 import { useBoard } from '../../../core/store/useBoardStore';
-import { type BoardProjection } from '../../../core/domain/BoardProjection';
+import { type BoardProjection } from '../../../core/domain/board/BoardProjection';
 
 interface NodeEntry {
   id: string;
