+++ src/ui/components/Slices/SliceInspectorView.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react';
 import { useBoard, useSlices, useSliceActions, useBoardStore } from '../../../core/store/useBoardStore';
-import { type BoardProjection } from '../../../core/domain/BoardProjection';
+import { type BoardProjection } from '../../../core/domain/board/BoardProjection';
 import { type VerticalSliceProjection, type ScenarioProjection } from '../../../core/domain/vertical-slice/VerticalSliceProjection';
 import { ScenarioDialog } from './ScenarioDialog';
